+++ CxAnalytixDaemon/Daemon.cs
@@ -9,6 +9,7 @@ using System.Reflection;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
+using CxAnalytix.TransformLogic.Interfaces;
 
 namespace CxAnalytixDaemon
 {
