+++ CxAnalytixService/CxAnalytixService.csproj
@@ -18,6 +18,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Configuration\Configuration.csproj" />
+    <ProjectReference Include="..\CxAuditTrails\CxAuditTrails.csproj" />
     <ProjectReference Include="..\CxRestClient\CxRestClient.csproj" />
     <ProjectReference Include="..\Log4NetOutput\Log4NetOutput.csproj" />
     <ProjectReference Include="..\MongoDBOutput\MongoDBOutput.csproj" />
