+++ Log4NetOutput/LoggerOut.cs
@@ -8,6 +8,7 @@ using System.Collections.Generic;
 using System.Reflection;
 using System.Threading;
 using System.Threading.Tasks;
+using CxAnalytix.TransformLogic.Interfaces;
 
 namespace CxAnalytix.Out.Log4NetOutput
 {
