+++ MongoDBOutput/MongoDBOut.cs
@@ -6,6 +6,7 @@ using MongoDB.Driver;
 using log4net;
 using CxAnalytix.Extensions;
 using System.Security.Cryptography;
+using CxAnalytix.TransformLogic.Interfaces;
 
 namespace CxAnalytix.Out.MongoDBOutput
 {
