+++ TransformLogic/TransformLogic.csproj
@@ -16,8 +16,11 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\CxAuditTrails\CxAuditTrails.csproj" />
     <ProjectReference Include="..\CxRestClient\CxRestClient.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="Interfaces\" />
   </ItemGroup>
 
 </Project>
